Extract renderSparkline into dedicated module
Improves code organization by separating sparkline rendering logic
from the main stats UI component.
